Extract VirtualTable to a separate module
